[maint] Moved error code into its own file for readability.
